requirements.txt
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
python-dotenv==1.2.1
requests==2.32.5
whitenoise==6.11.0
